move 'magic'-related stuff to dedicated file
