+++ src/databricks/labs/ucx/source_code/notebooks/sources.py
@@ -35,7 +35,7 @@ from databricks.labs.ucx.source_code.linters.imports import (
     SysPathChange,
     UnresolvedPath,
 )
-from databricks.labs.ucx.source_code.python.python_analyzer import MagicLine
+from databricks.labs.ucx.source_code.notebooks.magic import MagicLine
 from databricks.labs.ucx.source_code.python.python_ast import Tree, NodeBase
 from databricks.labs.ucx.source_code.notebooks.cells import (
     CellLanguage,
